Refactor: RedisService에 increment 메서드 추가 및 접근 로직 변경
