+++ src/main/java/ojosama/talkak/reaction/controller/ReactionController.java
@@ -1,7 +1,7 @@
 package ojosama.talkak.reaction.controller;
 
 
-import ojosama.talkak.redis.service.ReactionService;
+import ojosama.talkak.reaction.service.ReactionService;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
