+++ src/main/java/ojosama/talkak/reaction/service/ReactionService.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.redis.service;
+package ojosama.talkak.reaction.service;
 import ojosama.talkak.common.exception.TalKakException;
 import ojosama.talkak.common.exception.code.ReactionError;
 import ojosama.talkak.redis.HashConverter;
@@ -16,15 +16,12 @@ public class ReactionService {
 
     private final ReactionsRepository reactionsRepository;
     private final RedisService redisService;
-    private final RedisTemplate<String, Object> redisTemplate;
     private final HashConverter hashConverter;
 
     public ReactionService(ReactionsRepository reactionsRepository, RedisService redisService,
-        RedisTemplate<String, Object> redisTemplate,
         HashConverter hashConverter) {
         this.reactionsRepository = reactionsRepository;
         this.redisService = redisService;
-        this.redisTemplate = redisTemplate;
         this.hashConverter = hashConverter;
     }
 
@@ -48,13 +45,13 @@ public class ReactionService {
 
     public VideoInfo incrementViewCount(Long categoryId, Long videoId) {
         String key = VideoKey.VIDEO_INFO.generateKey(categoryId, videoId);
-        redisTemplate.opsForHash().increment(key, VideoHashKey.VIEW_COUNT.getKey(), 1);
+        redisService.incrementHashValue(key, VideoHashKey.VIEW_COUNT.getKey(), 1);
         return hashConverter.FromMap(redisService.getHashOps(key), VideoInfo.class);
     }
 
     public VideoInfo incrementLikeCount(Long categoryId, Long videoId) {
         String key = VideoKey.VIDEO_INFO.generateKey(categoryId, videoId);
-        redisTemplate.opsForHash().increment(key, VideoHashKey.LIKE_COUNT.getKey(), 1);
+        redisService.incrementHashValue(key, VideoHashKey.LIKE_COUNT.getKey(), 1);
         return hashConverter.FromMap(redisService.getHashOps(key), VideoInfo.class);
     }
